Implement cloud functions integration and API improvements
- Add cloud functions proxy for sports API calls
- Implement cloud API service for Flutter app
- Update Edge screen with improved UI and error handling
- Fix API gateway and sports API service configurations
- Add test files for cloud functions integration
- Update dependencies and configuration files

+++ bragging_rights_app/lib/services/settlement_service.dart
@@ -292,8 +292,8 @@ class SettlementService {
       for (int i = 0; i < winners.length && i < 3; i++) {
         final position = '${i + 1}';
         final prize = prizeStructure[position] ?? 0;
-        winners[i].payout = prize;
-        totalPayout += prize;
+        winners[i].payout = prize.toInt();
+        totalPayout += prize.toInt();
       }
     } else {
       // Equal distribution among winners
